test fixes for integer mode

+++ src/exercises/runway/funcs2.c
@@ -5,8 +5,8 @@
 struct State increment_Runway_Time(struct State s)
 /* --BEGIN-- */
 /*@ requires valid_state(s);
-             0i32 <= s.Runway_Time;
-             s.Runway_Time <= 4i32;
+             0 <= s.Runway_Time;
+             s.Runway_Time <= 4;
              s.ModeA == ACTIVE() || s.ModeD == ACTIVE();
     ensures  valid_state(return);
              s.Plane_Counter == return.Plane_Counter;
@@ -24,7 +24,7 @@ struct State reset_Runway_Time(struct State s)
 /* --BEGIN-- */
 /*@ requires valid_state(s);
     ensures  valid_state(return);
-             return.Runway_Time == 0i32;
+             return.Runway_Time == 0;
              s.ModeA == return.ModeA;
              s.ModeD == return.ModeD;
              s.W_A == return.W_A;
@@ -41,17 +41,17 @@ struct State reset_Runway_Time(struct State s)
 struct State arrive(struct State s)
 /* --BEGIN-- */
 /*@ requires valid_state(s);
-             s.ModeA == ACTIVE() && s.W_A >= 1i32;
-             s.Plane_Counter <= 2i32;
+             s.ModeA == ACTIVE() && s.W_A >= 1;
+             s.Plane_Counter <= 2;
     ensures  valid_state(return);
              s.Runway_Time == return.Runway_Time;
              s.ModeA == return.ModeA;
              s.ModeD == return.ModeD;
              s.W_D == return.W_D;
-             s.W_D == 0i32 
+             s.W_D == 0 
                implies s.Plane_Counter == return.Plane_Counter;
-             s.W_D > 0i32 
-               implies s.Plane_Counter == return.Plane_Counter - 1i32;
+             s.W_D > 0 
+               implies s.Plane_Counter == return.Plane_Counter - 1;
 @*/
 /* --END-- */
 {
@@ -66,8 +66,8 @@ struct State arrive(struct State s)
 struct State depart(struct State s)
 /* --BEGIN-- */
 /*@ requires valid_state(s);
-             s.ModeD == ACTIVE() && s.W_D >=1i32;
-             s.Plane_Counter <= 2i32;
+             s.ModeD == ACTIVE() && s.W_D >=1;
+             s.Plane_Counter <= 2;
     ensures  valid_state(return);
               s.Runway_Time == return.Runway_Time;
               s.ModeA == return.ModeA;
@@ -88,7 +88,7 @@ struct State switch_modes(struct State s)
 /* --BEGIN-- */
 /*@ requires valid_state(s);
              s.ModeA == ACTIVE() || s.ModeD == ACTIVE();
-             s.Plane_Counter == 0i32;
+             s.Plane_Counter == 0;
     ensures  valid_state(return);
              return.ModeA == ACTIVE() || return.ModeD == ACTIVE();
              return.ModeA == s.ModeD;
@@ -123,13 +123,13 @@ struct State switch_modes(struct State s)
 struct State tick(struct State s)
 /* --BEGIN-- */
 /*@ requires valid_state(s);
-             (i64) s.Plane_Counter < 2147483647i64;
-             (i64) s.W_A < 2147483647i64;
-             (i64) s.W_D < 2147483647i64;
+             s.Plane_Counter < 2147483647;
+             s.W_A < 2147483647;
+             s.W_D < 2147483647;
     ensures valid_state(return);
-            (s.W_A > 0i32 && s.W_D == 0i32 && s.Runway_Time == 0i32 
+            (s.W_A > 0 && s.W_D == 0 && s.Runway_Time == 0 
                implies return.ModeA == ACTIVE());
-            (s.W_D > 0i32 && s.W_A == 0i32 && s.Runway_Time == 0i32 
+            (s.W_D > 0 && s.W_A == 0 && s.Runway_Time == 0 
                implies return.ModeD == ACTIVE());
 @*/
 /* --END-- */
